requirements.txt
pandas
sentence-transformers
faiss-cpu
numpy
python-dotenv
groq
streamlit
huggingface_hub
fsspec
